requirements.txt
adal==0.4.7
asn1crypto==0.23.0
azure-common==1.1.8
-e git+git@github.com:energizedwork/azure-log-analytics-alerts-cli.git#egg=Azure_Log_Analytics_Alerts
azure-mgmt-loganalytics==0.1.0
azure-mgmt-nspkg==2.0.0
azure-nspkg==2.0.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.3
idna==2.6
isodate==0.6.0
keyring==10.4.0
msrest==0.4.18
msrestazure==0.4.16
oauthlib==2.0.6
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
requests==2.18.4
requests-oauthlib==0.8.0
SecretStorage==2.3.1
six==1.11.0
urllib3==1.22
